requirements.txt
byte>=0.0.0, <1.0.0
byte-sqlite>=0.0.0, <1.0.0
pysqlite>=2.8.0, <3.0.0
